Solution to Calculate and Print the Area of a Circle

+++ Calculate and Print the Area of a Circle@@6788d5012db25efec946f943/code.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 int main(){
-    int radius;
-    scanf("%d",radius);
-    printf("Area: %d",radius*radius*3.14);
+    float radius;
+    scanf("%f",radius);
+    printf("Area: %.2f",radius*radius*3.14);
     return 0;
 }
